# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface PatientRegisterRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
  dateOfBirth?: string;
  gender?: string;
  language?: string;
  region?: string;
}

export interface DoctorRegisterRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
  registrationNumber: string;
  fatherName: string;
  year: string;
  stateCouncil: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  userId: number;
  role: string;
  name: string;
  email: string;
  token?: string;
}

export interface DoctorRegisterResponse {
  userId: number;
  nmcVerified: boolean;
  verificationStatus: string;
  message: string;
}

export interface MeResponse {
  userId: number;
  role: string;
  name: string;
  email: string;
  verificationStatus?: string;
  isOnline?: boolean;
  profileComplete?: boolean;
  doctorId?: number;
  patientId?: number;
}

export interface DoctorProfileUpdateRequest {
  specialty?: string;
  subSpecialty?: string;
  languages?: string[];
  regions?: string[];
  experience?: number;
  education?: string;
  bio?: string;
  consultationFee?: number;
  symptoms?: string[];
}

export interface AvailabilityUpdateRequest {
  isOnline: boolean;
}

export interface DoctorProfile {
  id: number;
  userId: number;
  name: string;
  email?: string;
  phone?: string;
  registrationNumber?: string;
  stateCouncil?: string;
  specialty?: string;
  subSpecialty?: string;
  languages?: string[];
  regions?: string[];
  experience?: number;
  education?: string;
  bio?: string;
  consultationFee?: number;
  isOnline?: boolean;
  verificationStatus?: string;
  symptoms?: string[];
  rating?: number;
  totalConsultations?: number;
}

export interface DoctorSearchResult {
  id: number;
  userId?: number;
  name: string;
  specialty?: string;
  subSpecialty?: string;
  languages?: string[];
  regions?: string[];
  experience?: number;
  consultationFee?: number;
  isOnline?: boolean;
  rating?: number;
  totalConsultations?: number;
  isSameRegion?: boolean;
  isSameLanguage?: boolean;
}

export interface PatientProfile {
  id: number;
  userId: number;
  name: string;
  email: string;
  phone?: string;
  dateOfBirth?: string;
  gender?: string;
  language?: string;
  region?: string;
  bloodGroup?: string;
  allergies?: string;
  medicalHistory?: string;
}

export interface PatientProfileUpdateRequest {
  name?: string;
  phone?: string;
  dateOfBirth?: string;
  gender?: string;
  language?: string;
  region?: string;
  bloodGroup?: string;
  allergies?: string;
  medicalHistory?: string;
}

export interface ConsultationRequestCreate {
  doctorId: number;
  symptoms: string;
  notes?: string;
}

export type ConsultationResponseRequestStatus =
  (typeof ConsultationResponseRequestStatus)[keyof typeof ConsultationResponseRequestStatus];

export const ConsultationResponseRequestStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface ConsultationResponseRequest {
  status: ConsultationResponseRequestStatus;
  reason?: string;
}

export interface ConsultationRequest {
  id: number;
  patientId: number;
  doctorId: number;
  symptoms: string;
  notes?: string;
  status: string;
  reason?: string;
  createdAt?: string;
  patientName?: string;
  patientPhone?: string;
  patientLanguage?: string;
  patientRegion?: string;
  patientMedicalHistory?: string;
  patientAllergies?: string;
  previousConsultations?: number;
}

export interface ConsultationSummary {
  id: number;
  doctorId?: number;
  doctorName?: string;
  specialty?: string;
  symptoms?: string;
  status: string;
  paymentStatus?: string;
  createdAt?: string;
  completedAt?: string;
}

export interface Prescription {
  id: number;
  consultationId: number;
  content: string;
  documentUrl?: string;
  notes?: string;
  createdAt?: string;
}

export type MedicationOrderDeliveryStatus =
  (typeof MedicationOrderDeliveryStatus)[keyof typeof MedicationOrderDeliveryStatus];

export const MedicationOrderDeliveryStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
} as const;

export interface MedicationOrder {
  id: number;
  consultationId: number;
  medications: string[];
  instructions?: string;
  pharmacyName?: string;
  deliveryStatus: MedicationOrderDeliveryStatus;
  createdAt?: string;
  updatedAt?: string;
}

export interface ConsultationDetail {
  id: number;
  patientId?: number;
  doctorId?: number;
  symptoms?: string;
  notes?: string;
  status: string;
  paymentStatus?: string;
  roomId?: string;
  createdAt?: string;
  completedAt?: string;
  doctor?: DoctorProfile;
  patient?: PatientProfile;
  prescriptions?: Prescription[];
  medications?: MedicationOrder[];
}

export interface PaymentResponse {
  consultationId: number;
  amount: number;
  currency: string;
  sessionId?: string;
  message?: string;
}

export interface PrescriptionCreate {
  content: string;
  documentUrl?: string;
  notes?: string;
}

export interface MedicationCreate {
  medications: string[];
  instructions?: string;
  pharmacyName?: string;
}

export type MedicationStatusUpdateDeliveryStatus =
  (typeof MedicationStatusUpdateDeliveryStatus)[keyof typeof MedicationStatusUpdateDeliveryStatus];

export const MedicationStatusUpdateDeliveryStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
} as const;

export interface MedicationStatusUpdate {
  deliveryStatus: MedicationStatusUpdateDeliveryStatus;
}

export interface MedicalRecord {
  id: number;
  consultationId: number;
  doctorName?: string;
  specialty?: string;
  symptoms?: string;
  prescriptions?: Prescription[];
  medications?: MedicationOrder[];
  date?: string;
}

export interface PendingDoctor {
  id: number;
  userId?: number;
  name: string;
  email?: string;
  phone?: string;
  registrationNumber: string;
  fatherName?: string;
  year?: string;
  stateCouncil?: string;
  nmcVerified?: boolean;
  verificationStatus: string;
  createdAt?: string;
}

export interface RejectRequest {
  reason: string;
}

export interface AdminStats {
  totalDoctors: number;
  pendingVerifications: number;
  totalPatients: number;
  totalConsultations: number;
  activeConsultations: number;
}

export type WebRtcTokenRequestRole =
  (typeof WebRtcTokenRequestRole)[keyof typeof WebRtcTokenRequestRole];

export const WebRtcTokenRequestRole = {
  doctor: "doctor",
  patient: "patient",
} as const;

export interface WebRtcTokenRequest {
  consultationId: number;
  role: WebRtcTokenRequestRole;
}

export interface WebRtcTokenResponse {
  roomId: string;
  token: string;
  appId?: string;
}

export type SearchDoctorsParams = {
  symptom?: string;
  specialty?: string;
  language?: string;
  region?: string;
  onlineOnly?: boolean;
};