# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/patient:
    post:
      operationId: registerPatient
      tags: [auth]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRegisterRequest"
      responses:
        "201":
          description: Patient registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/doctor:
    post:
      operationId: registerDoctor
      tags: [auth]
      summary: Register a new doctor and verify with NMC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRegisterRequest"
      responses:
        "201":
          description: Doctor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorRegisterResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/doctors/{id}/approve:
    put:
      operationId: approveDoctor
      tags: [admin]
      summary: Approve a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/doctors/{id}/reject:
    put:
      operationId: rejectDoctor
      tags: [admin]
      summary: Reject a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Doctor rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PatientRegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        language:
          type: string
        region:
          type: string
      required:
        - name
        - email
        - password
    DoctorRegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        registrationNumber:
          type: string
        fatherName:
          type: string
        year:
          type: string
        stateCouncil:
          type: string
      required:
        - name
        - email
        - password
        - registrationNumber
        - fatherName
        - year
        - stateCouncil
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
        name:
          type: string
        email:
          type: string
        token:
          type: string
      required:
        - userId
        - role
        - name
        - email
    DoctorRegisterResponse:
      type: object
      properties:
        userId:
          type: integer
        nmcVerified:
          type: boolean
        verificationStatus:
          type: string
        message:
          type: string
      required:
        - userId
        - nmcVerified
        - verificationStatus
        - message
    MeResponse:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
        name:
          type: string
        email:
          type: string
        verificationStatus:
          type: string
        isOnline:
          type: boolean
        profileComplete:
          type: boolean
        doctorId:
          type: integer
        patientId:
          type: integer
      required:
        - userId
        - role
        - name
        - email
...
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...